+++ vulkano-shaders/src/structs.rs
@@ -9,55 +9,38 @@
 
 use std::mem;
 
-use enums;
-use parse;
+use syn::Ident;
+use proc_macro2::{Span, TokenStream};
+
+use parse::{Instruction, Spirv};
+use enums::Decoration;
+use spirv_search;
 
 /// Translates all the structs that are contained in the SPIR-V document as Rust structs.
-pub fn write_structs(doc: &parse::Spirv) -> String {
-    let mut result = String::new();
-
+pub fn write_structs(doc: &Spirv) -> TokenStream {
+    let mut structs = vec!();
     for instruction in &doc.instructions {
         match *instruction {
-            parse::Instruction::TypeStruct {
-                result_id,
-                ref member_types,
-            } => {
-                let (s, _) = write_struct(doc, result_id, member_types);
-                result.push_str(&s);
-                result.push_str("\n");
-            },
-            _ => (),
+            Instruction::TypeStruct { result_id, ref member_types } =>
+                structs.push(write_struct(doc, result_id, member_types).0),
+            _ => ()
         }
     }
 
-    result
-}
-
-/// Represents a rust struct member
-struct Member {
-    name: String,
-    value: String,
-    offset: Option<usize>,
-}
-
-impl Member {
-    fn declaration_text(&self) -> String {
-        let offset = match self.offset {
-            Some(o) => format!("/* offset: {} */", o),
-            _ => "".to_owned(),
-        };
-        format!("    pub {}: {} {}", self.name, self.value, offset)
-    }
-    fn copy_text(&self) -> String {
-        format!("            {name}: self.{name}", name = self.name)
+    quote!{
+        #( #structs )*
     }
 }
 
 /// Analyzes a single struct, returns a string containing its Rust definition, plus its size.
-fn write_struct(doc: &parse::Spirv, struct_id: u32, members: &[u32]) -> (String, Option<usize>) {
-    let name = ::name_from_id(doc, struct_id);
+fn write_struct(doc: &Spirv, struct_id: u32, members: &[u32]) -> (TokenStream, Option<usize>) {
+    let name = Ident::new(&spirv_search::name_from_id(doc, struct_id), Span::call_site());
 
     // The members of this struct.
+    struct Member {
+        pub name: Ident,
+        pub ty: TokenStream,
+    }
     let mut rust_members = Vec::with_capacity(members.len());
 
     // Padding structs will be named `_paddingN` where `N` is determined by this variable.
@@ -70,12 +53,12 @@ fn write_struct(doc: &parse::Spirv, struct_id: u32, members: &[u32]) -> (String,
     for (num, &member) in members.iter().enumerate() {
         // Compute infos about the member.
         let (ty, rust_size, rust_align) = type_from_id(doc, member);
-        let member_name = ::member_name_from_id(doc, struct_id, num as u32);
+        let member_name = spirv_search::member_name_from_id(doc, struct_id, num as u32);
 
         // Ignore the whole struct is a member is built in, which includes
         // `gl_Position` for example.
         if is_builtin_member(doc, struct_id, num as u32) {
-            return (String::new(), None); // TODO: is this correct? shouldn't it return a correct struct but with a flag or something?
+            return (quote!{}, None); // TODO: is this correct? shouldn't it return a correct struct but with a flag or something?
         }
 
         // Finding offset of the current member, as requested by the SPIR-V code.
@@ -83,10 +66,10 @@ fn write_struct(doc: &parse::Spirv, struct_id: u32, members: &[u32]) -> (String,
             .iter()
             .filter_map(|i| {
                 match *i {
-                    parse::Instruction::MemberDecorate {
+                    Instruction::MemberDecorate {
                         target_id,
                         member,
-                        decoration: enums::Decoration::DecorationOffset,
+                        decoration: Decoration::DecorationOffset,
                         ref params,
                     } if target_id == struct_id && member as usize == num => {
                         return Some(params[0]);
@@ -102,7 +85,7 @@ fn write_struct(doc: &parse::Spirv, struct_id: u32, members: &[u32]) -> (String,
         // variables only. Ignoring these.
         let spirv_offset = match spirv_offset {
             Some(o) => o as usize,
-            None => return (String::new(), None),        // TODO: shouldn't we return and let the caller ignore it instead?
+            None => return (quote!{}, None), // TODO: shouldn't we return and let the caller ignore it instead?
         };
 
         // We need to add a dummy field if necessary.
@@ -124,10 +107,9 @@ fn write_struct(doc: &parse::Spirv, struct_id: u32, members: &[u32]) -> (String,
                 let padding_num = next_padding_num;
                 next_padding_num += 1;
                 rust_members.push(Member {
-                                      name: format!("_dummy{}", padding_num),
-                                      value: format!("[u8; {}]", diff),
-                                      offset: None,
-                                  });
+                    name: Ident::new(&format!("_dummy{}", padding_num), Span::call_site()),
+                    ty: quote!{ [u8; #diff] },
+                });
                 *current_rust_offset += diff;
             }
         }
@@ -140,29 +122,28 @@ fn write_struct(doc: &parse::Spirv, struct_id: u32, members: &[u32]) -> (String,
         }
 
         rust_members.push(Member {
-                              name: member_name.to_owned(),
-                              value: ty,
-                              offset: Some(spirv_offset),
-                          });
+            name: Ident::new(&member_name, Span::call_site()),
+            ty,
+        });
     }
 
     // Try determine the total size of the struct in order to add padding at the end of the struct.
     let spirv_req_total_size = doc.instructions
         .iter()
         .filter_map(|i| match *i {
-                        parse::Instruction::Decorate {
+                        Instruction::Decorate {
                             target_id,
-                            decoration: enums::Decoration::DecorationArrayStride,
+                            decoration: Decoration::DecorationArrayStride,
                             ref params,
                         } => {
                             for inst in doc.instructions.iter() {
                                 match *inst {
-                                    parse::Instruction::TypeArray {
+                                    Instruction::TypeArray {
                                         result_id, type_id, ..
                                     } if result_id == target_id && type_id == struct_id => {
                                         return Some(params[0]);
                                     },
-                                    parse::Instruction::TypeRuntimeArray { result_id, type_id }
+                                    Instruction::TypeRuntimeArray { result_id, type_id }
                                         if result_id == target_id && type_id == struct_id => {
                                         return Some(params[0]);
                                     },
@@ -186,58 +167,63 @@ fn write_struct(doc: &parse::Spirv, struct_id: u32, members: &[u32]) -> (String,
         let diff = req_size.checked_sub(cur_size as u32).unwrap();
         if diff >= 1 {
             rust_members.push(Member {
-                                  name: format!("_dummy{}", next_padding_num),
-                                  value: format!("[u8; {}]", diff),
-                                  offset: None,
-                              });
+                name: Ident::new(&format!("_dummy{}", next_padding_num), Span::call_site()),
+                ty: quote!{ [u8; {}] },
+            });
         }
     }
 
     // We can only implement Clone if there's no unsized member in the struct.
-    let (impl_text, derive_text) = if current_rust_offset.is_some() {
-        let i = format!("\nimpl Clone for {name} {{\n    fn clone(&self) -> Self {{\n        \
-                         {name} {{\n{copies}\n        }}\n    }}\n}}\n",
-                        name = name,
-                        copies = rust_members
-                            .iter()
-                            .map(Member::copy_text)
-                            .collect::<Vec<_>>()
-                            .join(",\n"));
-        (i, "#[derive(Copy)]")
+    let (clone_impl, copy_derive) = if current_rust_offset.is_some() {
+        let mut copies = vec!();
+        for member in &rust_members {
+            let name = &member.name;
+            copies.push(quote!{ #name: self.#name, });
+        }
+        (
+            // Clone is implemented manually because members can be large arrays
+            // that do not implement Clone, but do implement Copy
+            quote!{
+                impl Clone for #name {
+                    fn clone(&self) -> Self {
+                        #name {
+                            #( #copies )*
+                        }
+                    }
+                }
+            },
+            quote!{ #[derive(Copy)] }
+        )
     } else {
-        ("".to_owned(), "")
+        (quote!{}, quote!{})
     };
 
-    let s =
-        format!("#[repr(C)]\n{derive_text}\n#[allow(non_snake_case)]\npub struct {name} \
-                 {{\n{members}\n}} /* total_size: {t:?} */\n{impl_text}",
-                name = name,
-                members = rust_members
-                    .iter()
-                    .map(Member::declaration_text)
-                    .collect::<Vec<_>>()
-                    .join(",\n"),
-                t = spirv_req_total_size,
-                impl_text = impl_text,
-                derive_text = derive_text);
-    (s,
-     spirv_req_total_size
-         .map(|sz| sz as usize)
-         .or(current_rust_offset))
+    let mut members = vec!();
+    for member in &rust_members {
+        let name = &member.name;
+        let ty = &member.ty;
+        members.push(quote!(pub #name: #ty,));
+    }
+
+    let ast = quote! {
+        #[repr(C)]
+        #copy_derive
+        #[allow(non_snake_case)]
+        pub struct #name {
+            #( #members )*
+        }
+        #clone_impl
+    };
+
+    (ast, spirv_req_total_size.map(|sz| sz as usize).or(current_rust_offset))
 }
 
 /// Returns true if a `BuiltIn` decorator is applied on a struct member.
-fn is_builtin_member(doc: &parse::Spirv, id: u32, member_id: u32) -> bool {
+fn is_builtin_member(doc: &Spirv, id: u32, member_id: u32) -> bool {
     for instruction in &doc.instructions {
         match *instruction {
-            parse::Instruction::MemberDecorate {
-                target_id,
-                member,
-                decoration: enums::Decoration::DecorationBuiltIn,
-                ..
-            } if target_id == id && member == member_id => {
-                return true;
-            },
+            Instruction::MemberDecorate { target_id, member, decoration: Decoration::DecorationBuiltIn, .. }
+                if target_id == id && member == member_id => { return true }
             _ => (),
         }
     }
@@ -248,17 +234,13 @@ fn is_builtin_member(doc: &parse::Spirv, id: u32, member_id: u32) -> bool {
 /// Returns the type name to put in the Rust struct, and its size and alignment.
 ///
 /// The size can be `None` if it's only known at runtime.
-pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>, usize) {
+pub fn type_from_id(doc: &Spirv, searched: u32) -> (TokenStream, Option<usize>, usize) {
     for instruction in doc.instructions.iter() {
         match instruction {
-            &parse::Instruction::TypeBool { result_id } if result_id == searched => {
+            &Instruction::TypeBool { result_id } if result_id == searched => {
                 panic!("Can't put booleans in structs")
-            },
-            &parse::Instruction::TypeInt {
-                result_id,
-                width,
-                signedness,
-            } if result_id == searched => {
+            }
+            &Instruction::TypeInt { result_id, width, signedness } if result_id == searched => {
                 match (width, signedness) {
                     (8, true) => {
                         #[repr(C)]
@@ -267,7 +249,7 @@ pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>
                             after: u8,
                         }
                         let size = unsafe { (&(&*(0 as *const Foo)).after) as *const u8 as usize };
-                        return ("i8".to_owned(), Some(size), mem::align_of::<Foo>());
+                        return (quote!{i8}, Some(size), mem::align_of::<Foo>());
                     },
                     (8, false) => {
                         #[repr(C)]
@@ -276,7 +258,7 @@ pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>
                             after: u8,
                         }
                         let size = unsafe { (&(&*(0 as *const Foo)).after) as *const u8 as usize };
-                        return ("u8".to_owned(), Some(size), mem::align_of::<Foo>());
+                        return (quote!{u8}, Some(size), mem::align_of::<Foo>());
                     },
                     (16, true) => {
                         #[repr(C)]
@@ -285,7 +267,7 @@ pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>
                             after: u8,
                         }
                         let size = unsafe { (&(&*(0 as *const Foo)).after) as *const u8 as usize };
-                        return ("i16".to_owned(), Some(size), mem::align_of::<Foo>());
+                        return (quote!{i16}, Some(size), mem::align_of::<Foo>());
                     },
                     (16, false) => {
                         #[repr(C)]
@@ -294,7 +276,7 @@ pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>
                             after: u8,
                         }
                         let size = unsafe { (&(&*(0 as *const Foo)).after) as *const u8 as usize };
-                        return ("u16".to_owned(), Some(size), mem::align_of::<Foo>());
+                        return (quote!{u16}, Some(size), mem::align_of::<Foo>());
                     },
                     (32, true) => {
                         #[repr(C)]
@@ -303,7 +285,7 @@ pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>
                             after: u8,
                         }
                         let size = unsafe { (&(&*(0 as *const Foo)).after) as *const u8 as usize };
-                        return ("i32".to_owned(), Some(size), mem::align_of::<Foo>());
+                        return (quote!{i32}, Some(size), mem::align_of::<Foo>());
                     },
                     (32, false) => {
                         #[repr(C)]
@@ -312,7 +294,7 @@ pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>
                             after: u8,
                         }
                         let size = unsafe { (&(&*(0 as *const Foo)).after) as *const u8 as usize };
-                        return ("u32".to_owned(), Some(size), mem::align_of::<Foo>());
+                        return (quote!{u32}, Some(size), mem::align_of::<Foo>());
                     },
                     (64, true) => {
                         #[repr(C)]
@@ -321,7 +303,7 @@ pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>
                             after: u8,
                         }
                         let size = unsafe { (&(&*(0 as *const Foo)).after) as *const u8 as usize };
-                        return ("i64".to_owned(), Some(size), mem::align_of::<Foo>());
+                        return (quote!{i64}, Some(size), mem::align_of::<Foo>());
                     },
                     (64, false) => {
                         #[repr(C)]
@@ -330,12 +312,12 @@ pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>
                             after: u8,
                         }
                         let size = unsafe { (&(&*(0 as *const Foo)).after) as *const u8 as usize };
-                        return ("u64".to_owned(), Some(size), mem::align_of::<Foo>());
+                        return (quote!{u64}, Some(size), mem::align_of::<Foo>());
                     },
                     _ => panic!("No Rust equivalent for an integer of width {}", width),
                 }
-            },
-            &parse::Instruction::TypeFloat { result_id, width } if result_id == searched => {
+            }
+            &Instruction::TypeFloat { result_id, width } if result_id == searched => {
                 match width {
                     32 => {
                         #[repr(C)]
@@ -344,7 +326,7 @@ pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>
                             after: u8,
                         }
                         let size = unsafe { (&(&*(0 as *const Foo)).after) as *const u8 as usize };
-                        return ("f32".to_owned(), Some(size), mem::align_of::<Foo>());
+                        return (quote!{f32}, Some(size), mem::align_of::<Foo>());
                     },
                     64 => {
                         #[repr(C)]
@@ -353,61 +335,58 @@ pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>
                             after: u8,
                         }
                         let size = unsafe { (&(&*(0 as *const Foo)).after) as *const u8 as usize };
-                        return ("f64".to_owned(), Some(size), mem::align_of::<Foo>());
+                        return (quote!{f64}, Some(size), mem::align_of::<Foo>());
                     },
                     _ => panic!("No Rust equivalent for a floating-point of width {}", width),
                 }
-            },
-            &parse::Instruction::TypeVector {
+            }
+            &Instruction::TypeVector {
                 result_id,
                 component_id,
                 count,
             } if result_id == searched => {
                 debug_assert_eq!(mem::align_of::<[u32; 3]>(), mem::align_of::<u32>());
-                let (t, t_size, t_align) = type_from_id(doc, component_id);
-                return (format!("[{}; {}]", t, count), t_size.map(|s| s * count as usize), t_align);
-            },
-            &parse::Instruction::TypeMatrix {
+                let (ty, t_size, t_align) = type_from_id(doc, component_id);
+                let array_length = count as usize;
+                let size = t_size.map(|s| s * count as usize);
+                return (quote!{ [#ty; #array_length] }, size, t_align);
+            }
+            &Instruction::TypeMatrix {
                 result_id,
                 column_type_id,
                 column_count,
             } if result_id == searched => {
                 // FIXME: row-major or column-major
                 debug_assert_eq!(mem::align_of::<[u32; 3]>(), mem::align_of::<u32>());
-                let (t, t_size, t_align) = type_from_id(doc, column_type_id);
-                return (format!("[{}; {}]", t, column_count),
-                        t_size.map(|s| s * column_count as usize),
-                        t_align);
-            },
-            &parse::Instruction::TypeArray {
+                let (ty, t_size, t_align) = type_from_id(doc, column_type_id);
+                let array_length = column_count as usize;
+                let size = t_size.map(|s| s * column_count as usize);
+                return (quote!{ [#ty; #array_length] }, size, t_align);
+            }
+            &Instruction::TypeArray {
                 result_id,
                 type_id,
                 length_id,
             } if result_id == searched => {
                 debug_assert_eq!(mem::align_of::<[u32; 3]>(), mem::align_of::<u32>());
-                let (t, t_size, t_align) = type_from_id(doc, type_id);
+                let (ty, t_size, t_align) = type_from_id(doc, type_id);
                 let t_size = t_size.expect("array components must be sized");
                 let len = doc.instructions
                     .iter()
                     .filter_map(|e| match e {
-                                    &parse::Instruction::Constant {
-                                        result_id,
-                                        ref data,
-                                        ..
-                                    } if result_id == length_id => Some(data.clone()),
-                                    _ => None,
-                                })
+                        &Instruction::Constant { result_id, ref data, .. }
+                            if result_id == length_id => Some(data.clone()),
+                        _ => None,
+                    })
                     .next()
                     .expect("failed to find array length");
                 let len = len.iter().rev().fold(0u64, |a, &b| (a << 32) | b as u64);
                 let stride = doc.instructions.iter().filter_map(|e| match e {
-                    parse::Instruction::Decorate{
+                    Instruction::Decorate {
                         target_id,
-                        decoration: enums::Decoration::DecorationArrayStride,
+                        decoration: Decoration::DecorationArrayStride,
                         ref params,
-                    } if *target_id == searched => {
-                        Some(params[0])
-                    },
+                    } if *target_id == searched => Some(params[0]),
                     _ => None,
                 })
                 .next().expect("failed to find ArrayStride decoration");
@@ -417,27 +396,30 @@ pub fn type_from_id(doc: &parse::Spirv, searched: u32) -> (String, Option<usize>
                             the array element in a struct or rounding up the size of a vector or matrix \
                             (e.g. increase a vec3 to a vec4)")
                 }
-                return (format!("[{}; {}]", t, len), Some(t_size * len as usize), t_align);
-            },
-            &parse::Instruction::TypeRuntimeArray { result_id, type_id }
+                let array_length = len as usize;
+                let size = Some(t_size * len as usize);
+                return (quote!{ [#ty; #array_length] }, size, t_align);
+            }
+            &Instruction::TypeRuntimeArray { result_id, type_id }
                 if result_id == searched => {
                 debug_assert_eq!(mem::align_of::<[u32; 3]>(), mem::align_of::<u32>());
-                let (t, _, t_align) = type_from_id(doc, type_id);
-                return (format!("[{}]", t), None, t_align);
-            },
-            &parse::Instruction::TypeStruct {
+                let (ty, _, t_align) = type_from_id(doc, type_id);
+                return (quote!{ [#ty] }, None, t_align);
+            }
+            &Instruction::TypeStruct {
                 result_id,
                 ref member_types,
             } if result_id == searched => {
                 // TODO: take the Offset member decorate into account?
-                let name = ::name_from_id(doc, result_id);
+                let name = Ident::new(&spirv_search::name_from_id(doc, result_id), Span::call_site());
+                let ty = quote!{ #name };
                 let (_, size) = write_struct(doc, result_id, member_types);
                 let align = member_types
                     .iter()
                     .map(|&t| type_from_id(doc, t).2)
                     .max()
                     .unwrap_or(1);
-                return (name, size, align);
+                return (ty, size, align);
             },
             _ => (),
         }